Add files via upload
The system automatically selects a number between a particular range and helps us guess the number. We enter our guesses and the system compares our input value with the selected number and tells us whether the selected number is greater than or equal to or smaller than our guessed number. There is no limit to the number of guesses that we can do until we guess the right number.